added JPA text class and updated the Integer ID to Long to reflect update since lesson was filmed

+++ src/main/java/com/dmartinez/customer/Customer.java
@@ -25,7 +25,7 @@ public class Customer {
             strategy = GenerationType.SEQUENCE,
             generator = "customer_id_seq"
     )
-    private Integer id;
+    private Long id;
     @Column(nullable = false)
     private String name;
     @Column(nullable = false)
@@ -41,18 +41,18 @@ public class Customer {
         this.age = age;
     }
 
-    public Customer(Integer id, String name, String email, Integer age) {
+    public Customer(Long id, String name, String email, Integer age) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.age = age;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
